Create code
A price comparison website act as the medium between
customers and retailers so customers can make purchase online for certain
retailers that provide such services. On the other words, the price comparison
website also has the role to promote the retailer/shop/hypermarket/supermarket
to the customers. The pressures on time and money especially in the current
economic situation where the living cost increases and there’s only little time to
do some shopping for household, a price comparison website will absolutely
become a great help towards consumers. Besides, users nowadays are very
comfortable with the internet that it has grown a wider variety of applications
from networking and now provide various references for the users.
XAMMP is a free and open-source cross-platform web server solution stack
package developed by Apache Friends, consisting mainly of the Apache HTTP
Server, MarinaDB database and interpreters for script written in the PHP
programming language.
The tools that will use for this system are:

• HTML as the main development platform. This HTML
will connected with PHP language
• MySQL as database platform
• Another language such as CSS and javaScript